refactor: reuse code moved in kyverno repo
