+++ backend/go.mod
@@ -8,7 +8,7 @@ require (
 	github.com/gin-contrib/cors v1.4.0
 	github.com/gin-gonic/gin v1.9.1
 	github.com/go-logr/logr v1.2.4
-	github.com/kyverno/kyverno v1.11.0-beta.1
+	github.com/kyverno/kyverno v1.11.0-beta.1.0.20230921175422-81f2646963bf
 	github.com/loopfz/gadgeto v0.11.3
 	github.com/spf13/cobra v1.7.0
 	github.com/stretchr/testify v1.8.4
