+++ backend/pkg/resource/helpers.go
@@ -1,13 +1,13 @@
-package loader
+package resource
 
 import (
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/resource/convert"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/resource/loader"
 	yamlutils "github.com/kyverno/kyverno/pkg/utils/yaml"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-
-	"github.com/kyverno/playground/backend/pkg/resource/convert"
 )
 
-func Load[T any](l Loader, content []byte) (*T, error) {
+func Load[T any](l loader.Loader, content []byte) (*T, error) {
 	untyped, err := l.Load(content)
 	if err != nil {
 		return nil, err
@@ -19,7 +19,7 @@ func Load[T any](l Loader, content []byte) (*T, error) {
 	return result, nil
 }
 
-func LoadResources(l Loader, content []byte) ([]unstructured.Unstructured, error) {
+func LoadResources(l loader.Loader, content []byte) ([]unstructured.Unstructured, error) {
 	documents, err := yamlutils.SplitDocuments(content)
 	if err != nil {
 		return nil, err
