+++ backend/pkg/utils/exception.go
@@ -2,13 +2,14 @@ package utils
 
 import (
 	kyvernov2alpha1 "github.com/kyverno/kyverno/api/kyverno/v2alpha1"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/resource/convert"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/resource/loader"
 
-	"github.com/kyverno/playground/backend/pkg/resource/convert"
-	"github.com/kyverno/playground/backend/pkg/resource/loader"
+	"github.com/kyverno/playground/backend/pkg/resource"
 )
 
 func LoadPolicyExceptions(l loader.Loader, content []byte) ([]*kyvernov2alpha1.PolicyException, error) {
-	untyped, err := loader.LoadResources(l, content)
+	untyped, err := resource.LoadResources(l, content)
 	if err != nil {
 		return nil, err
 	}
